refactor: 기본 구조 조정
도메인로직 분리를 위한 기본 구조 조정
