+++ src/main/java/lotto/Application.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.controller.LottoGameSimulator;
+
 public class Application {
     public static void main(String[] args) {
         LottoGameSimulator lottoGameSimulator = new LottoGameSimulator();
