--- src/main/java/lotto/Validator.java
@@ -1,4 +1,0 @@
-package lotto;
-
-public class Validator {
-}
