+++ src/main/java/lotto/domain/generator/WinningNumbersGenerator.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.domain.generator;
+
+import lotto.domain.WinningNumbers;
 
 import java.util.ArrayList;
 import java.util.Arrays;
