+++ src/main/java/lotto/domain/validator/Validator.java
@@ -1,0 +1,4 @@
+package lotto.domain.validator;
+
+public class Validator {
+}
