+++ src/main/java/lotto/domain/validator/WinningNumbersValidator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.validator;
 
 import java.util.ArrayList;
 import java.util.Arrays;
